replace-in-archive improvements..
